Be more generic with what file we accept
The basic struct was already rather abstract as it was only requiring seek +
read, but the implementation was restricted to `File`. This alleviates the
restriction to allow any kind of element that implements those two traits. This
is required for environments that cannot provide a file system file or where the
 file is already inside memory.