Properly propagate errors outwards
When trying to load a file that isn't a supported NetCDF file, the parse would
so far panic at an `unwrap()`. Obviously when a library would present a proper
error to the user this was not possible. This now propagates the error past that
 boundary so that calling `new()` or `new_from_file()` may result in an error if
 the file is not valid.

We can't pass a `NomError` outwards because it would generally outlive the
input, which would make the temporary buffer irrelevant. At the entry, we make
sure that we only return errors that are not nom-errors. We could make it so
that we split off the error into two, but we'd still have to create a union of
those two types at some point again anyway, resulting in the same type parameter
requirement and problems.

The String type param is a little unfortunate, but seems like an OK compromise
to be able to keep using the same error type for internal and external errors.